+++ src/BuildAcceptors.c
@@ -37,17 +37,17 @@ extern long NUMSITES;
 /* Additional profiles */
 extern int BP;
 extern int PPT;
-float ComputeU2BranchProfile(char    *s,
-                             long    positionAcc,
-                             long    limitRight,
-                             profile *p,
-                             site    *splicesite){
-    float maxScore;
-    float score;
-    int   index;
-    int   Opt;
-    long  end;
-    long  i, j;
+double ComputeU2BranchProfile(char    *s,
+                              long    positionAcc,
+                              long    limitRight,
+                              profile *p,
+                              site    *splicesite){
+    double maxScore;
+    double score;
+    int    index;
+    int    Opt;
+    long   end;
+    long   i, j;
 
     maxScore = -INF;
     /*      char mess[MAXSTRING];  */
@@ -73,8 +73,8 @@ float ComputeU2BranchProfile(char    *s,
             }
         }
 
-        score = score - p->penalty_factor * (((float) (labs(i + p->offset - Opt)) / ((float) (p->acc_context - p->offset - p->opt_dist))) * ((float) (labs(i + p->offset
-                                                                                                                                                         - Opt)) / ((float) (p->acc_context - p->offset - p->opt_dist))));
+        score = score - p->penalty_factor * (((double) (labs(i + p->offset - Opt)) / ((double) (p->acc_context - p->offset - p->opt_dist))) * ((double) (labs(i + p->offset
+                                                                                                                                                              - Opt)) / ((double) (p->acc_context - p->offset - p->opt_dist))));
         score = p->afactor + (p->bfactor * score);
 
         if ((score >= maxScore) && (score > p->cutoff)) {
@@ -90,16 +90,16 @@ float ComputeU2BranchProfile(char    *s,
     return maxScore;
 }
 
-float ComputeU2PPTProfile(char    *s,
-                          long    positionAcc,
-                          long    limitRight,
-                          profile *p,
-                          site    *splicesite){
-    float maxScore;
-    float score;
-    int   index;
-    long  end;
-    long  i, j;
+double ComputeU2PPTProfile(char    *s,
+                           long    positionAcc,
+                           long    limitRight,
+                           profile *p,
+                           site    *splicesite){
+    double maxScore;
+    double score;
+    int    index;
+    long   end;
+    long   i, j;
 
     maxScore = -INF;
 
@@ -149,17 +149,17 @@ long  BuildAcceptors(char    *s,
                      long    l2,
                      long    ns,
                      long    nsites){
-    int   i, j;
-    char  *sOriginal;
-    float score;
-    float scoreBP;
-    float scorePPT;
-    float scoreAcc;
+    int    i, j;
+    char   *sOriginal;
+    double score;
+    double scoreBP;
+    double scorePPT;
+    double scoreAcc;
     /*   long ns,is; */
-    long  is;
-    long  left, right;
-    int   index;
-    float cutoff;
+    long   is;
+    long   left, right;
+    int    index;
+    double cutoff;
 
     /* Final number of predicted signals (that type) */
     /*   ns = 0; */
